collector/requirements.txt
flask==3.0.0
flask-cors==4.0.0
vk-api==11.9.9
requests==2.31.0
psycopg2-binary==2.9.9
pymongo==4.6.1
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1